Unify agent server settings and extract e2e tests
